Remove old container and canary tests
